moving connection string parsing into it's own file
also adding full test coverage for it
